feat: always provide `Default` for `MultiCompiler` (#188)
Closes https://github.com/foundry-rs/compilers/issues/187
Closes https://github.com/foundry-rs/compilers/issues/168

Updates `MultiCompiler` and makes solc compiler on it optional. If
`svm-solc` is not activated, it is attempted to initialize solc via
`Solc::new("solc")`, and if solc binary is not available in command
line, it is set to None.

We could also add `svm-solc` to default features as I think people would
often prefer enabling it as it's required to emulate foundry's default
behavior.