feat: better artifact filenames for different profiles (#241)
Closes https://github.com/foundry-rs/foundry/issues/9722

This PR introduces a concept of "primary" profile. Primary profile is
the one which we've chosen for input source when analyzing it on its own
(i.e not as a dependency of other contract). That way `Counter.json`
would always correspond to an artifact compiled with the default profile
if it was possible.